Remove . and ~ from alias generator alphabet.
Heather thought that "." may be easy for people to miss when typing
the shortened URLs. While that's a niche case, it may help.

Removed "~" for the same use case, but because it's not a commonly used
character for most people. They may not be able to find it right away,
especially on phone keyboards.