Remove atdgen validation code from configs.
I shouldn't have committed it to begin with. It ended up not being
necessary since it was just to check file paths for real, readable
files. But since the logic changed to read those files at server
start, any issues would be reported anyway without the need for
this validation.